refactor(runtime): factor out code between build.rs and worker.rs
